--- src/modbus/dataframes/temperature_program.py
@@ -1,9 +1,0 @@
-from dataclasses import dataclass
-
-
-@dataclass
-class TemperatureProgram:
-    target_temperature: int
-    point_position: int
-    raising_time: int
-    holding_time: int
