+++ src/modbus/utils/enums/__init__.py
@@ -1,0 +1,1 @@
+from .operating_mode import OperatingMode
